Added Food Blobs
Ive made it so the map currently has 3 5x5 blobs of food. This is the
sizes for the tournament but the 11 blobs would take up too much room.
It is a little buggy if the map is too small, because the blobs don’t
fit.
Try and keep world sizes above 20x20 or reduce the number of blobs
within the method.